Remove readme (use $comment for status in current and changelog.md file instead) for old content.
The readme was originally added so additional information could
be added about schema in a standard file since JSON does not
allow comments. It included status of schema and change log.

The description in the current file should be enough for most
schema. A problem with using the readme for status is that
it will make automation that applies to [draft|completed|retired...]
schema more hassle. A changelog of some form needs to exist so
review of git history is not required by most writers.

This commit removes the readme from emptyj and emptyjref. To those 2
schema it added a value so the status can be found via loading
the current.json file. It needed to use a JSON schema keyword
added in draft7 else SRC would fail. The value and flag
in commit is: "$current": "status-draft" where the flag
includes a category so easier to filter sets of flags. File
changelog.md was copied from how wikimedia tracks how
schema have changed over years. A changelog for a release
has date and person/team doing the change in brackets