Use examples for folder name. Stage 2. Releases of all schemas using new path
A schema set or group containing a schema folders should never move
location within a schema root for lifetime of repo. However, is
is possible as this commit shows.

All $ref schema must be rebuilt BEFORE all schema using it. This
means usage of build-new is required on schema in a specific order
so knowledge of usage of $ref is useful before doing changes. For
example, in the reposhow group basic/fast and basic/slow schema
must be built before advfast and advslow. Secondly,
advslow must be built before advtop. Then SRC will pass.

This commit changes the zdemo schema set to become examples
schema set with new releases of all schema within. It updates
the title and $id in every current file with build-new on each
and SRC passes.