Added build-main-diff command so new/changed current.json build
This commit adds a new `npm run build-main-modified` command
which looks for current.json files that are new or have changed
since the last merge to main. It does this by using a git
diff against origin/main rather than a local copy of main
since this is required when run by github actions.